#98 refactor: DateFormatter 인스턴스가 매번 초기화되지 않도록 LifePoopDateFormatter 싱글톤 클래스 구현하여 성능 최적화
