+++ Projects/Utils/Sources/Extension/Date+localizedString.swift
@@ -9,27 +9,11 @@
 import Foundation
 
 public extension Date {
-    private var userLocale: Locale {
-        if let regionCode = Locale.current.regionCode,
-           let languageCode = Locale.preferredLanguages.first?.components(separatedBy: "-").first {
-            let localeIdentifier = "\(languageCode)_\(regionCode)"
-            return Locale(identifier: localeIdentifier)
-        } else {
-            return Locale.current
-        }
-    }
-    
-    var localizedString: String {
-        let formatter = DateFormatter()
-        formatter.dateStyle = .long
-        formatter.locale = userLocale
-        return formatter.string(from: self)
+    var localizedDateString: String {
+        return LifePoopDateFormatter.shared.localizedString(from: self, dateStyle: .long, timeStyle: .none)
     }
     
     var localizedTimeString: String {
-        let formatter = DateFormatter()
-        formatter.timeStyle = .short
-        formatter.locale = userLocale
-        return formatter.string(from: self)
+        return LifePoopDateFormatter.shared.localizedString(from: self, dateStyle: .none, timeStyle: .short)
     }
 }
